Clear the four lint errors that stood behind the first four
`just check` runs `cargo clippy --workspace --all-targets -- -D warnings`.
Clippy stops a crate at its first failing target, so the four errors cleared in
c647262c were hiding four more behind them. Each fix revealed the next.

    crates/cachette-core/tests/military_strength.rs:226           useless_conversion
    crates/cachette-core/tests/fog_last_seen.rs:81                useless_conversion
    crates/cachette-core/tests/a_faction_reads_one_flat_array.rs:367  useless_conversion
    crates/cachette-core/examples/weather_band_probe.rs:147       manual_range_contains

`SightRules::ceiling` already answers `u32`, so three call sites wrapped it in
`u32::from` for nothing. The belt probe tested a latitude against two bounds
with a disjunction.

Every one of the four is semantics-preserving. Three came from `cargo clippy
--fix --workspace --all-targets`, which only applies machine-applicable
suggestions, and the fourth was the same conversion by hand. The diff is four
lines in four files and was read before it was staged.

    cargo clippy --workspace --all-targets -- -D warnings     clean

None of these eight lint errors comes from the weather work in this session.
They stand on a clean tree at 21aa3fc0.

+++ crates/cachette-core/tests/fog_last_seen.rs
@@ -78,7 +78,7 @@ fn two_far_cells(world: &World) -> (Axial, Axial) {
     let near = dry[0];
     let far = *dry.last().expect("the fixture world holds passable ground");
     assert!(
-        near.distance(far) > u32::from(world.sight_rules().ceiling()) * 2,
+        near.distance(far) > world.sight_rules().ceiling() * 2,
         "the two addresses must share no sight at any admitted radius"
     );
     assert_ne!(
